pyproject.toml
[project]
name = "memory-cli"
version = "0.1.0"
description = "A personal media manager for photos, videos, etc."
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    "click>=8.0.0",
    "Pillow>=10.0.0", # For image metadata (EXIF)
]
readme = "README.md"
requires-python = ">=3.9"

[project.scripts]
memory = "memory.cli:cli"

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
